FULL navbar, construction page, & style (issue #3)
Add navigation bar with hover color-changing and links to a new page
called under-construction-notice.html which just says the page is
still under construction. The home button links to the landing page.
Also add the fonts agreed upon and some spacing for the nav bar and
general page. ("Full commit" because not all files were added on the
last commit.)