Adds editor page + function (issue #37)
Adds the editor.php page which is similar to the creator.php except it
displays with information already filled in about your little guy, which can
then be modified and resubmitted. The editguyhelper.php then modifies
the database to update the listing for that guy's id. Also, renamed the
helper files to clarify them from the creator/editor.