Created basic files for the app + nav
Created the basic files for the app version, and implemented the
Navigation library. Created the main App file, which is just the
container, and then the Create, Edit, and Home files, to be filled in
with their actual content later. They only have basic buttons and text
as of now.