Created homeScreen layout
For Issue #68.
Added LittleGuy display to the homeScreen, for both your little guys and
other people's. They are not yet connected to the database, but display
everything. It's separated into LittleGuy components who display a row
of information about themselves, and various headers and styling. Also
added the pictures of the variants to display accordingly.